// frontend/src/types/taskStatus.ts
export enum TaskStatusEnum {
  PENDING = "PENDING",
  DOWNLOADING = "DOWNLOADING",
  PROCESSING_INVENTORY = "PROCESSING_INVENTORY",
  PROCESSING_SLA = "PROCESSING_SLA",
  PROCESSING_INCIDENTS = "PROCESSING_INCIDENTS",
  GENERATING_CHARTS = "GENERATING_CHARTS",
  GENERATING_PPT = "GENERATING_PPT",
  CLEANUP = "CLEANUP",
  COMPLETED = "COMPLETED",
  FAILED = "FAILED",
}

export interface TaskStatus {
  value: TaskStatusEnum;
  label: string;
  description: string;
}